Add an enum-in-range check.
This check can be enabled by using the flag --enum-range-check in
goto-instrument. It inserts assertions before each use of any enum
type expressions that the expression's value is one of the valid
enum values.

Note: It doesn't insert checks for occurrences of enum expressions
on the left hand side of assignments.