Add STL main entry point support
Expands the current entry point selection by an additional check for a
block with the name "Main". Note: This does not reflect the typical
behaviour of Siemens TIA as the main block can have any name. In order
to find the 'true' main block it is necessary to parse block properties,
which is not possible yet. The check for "Main" is intended to be the
last resort if no entry point is specified (or if the true main block
could not be resolved in the future).