Make goto_rw less conservative
When handling the case of pointers, goto_rw would mark both the object pointed
to and the pointer itself as written. For example, if x = &i, the line `*x = 1'
would yield a write set containing both x and i, instead of just i.

This resolves that issue by ensuring that value_set_dereference always gives at
least one object that the dereference can refer to.

This also resolves a bug wherein &a is marked as reading from a by removing the
ID_symbol special case from get_objects_address_of.