Add field-sensitivity tests
These look at various cases of struct-field assignment, directly and via pointers; only one
case is identified as not using the field-sensitivity machinery so far (that of a pointer that
may point to several different offsets within the same target allocation).

These are added now so that forthcoming changes to field-sensitivity have a reasonably comprehensive
regression test suite to check against.