Initial commit of symex complexity work
This work allows for a more intelligent way of pathing  around / slicing
away problematic code dynamically as symex runs. Every path will have a
'complexity' rating that shows currently how difficult the code it's
running is for symex to analyze (alternatively also for the SAT solver
to solve) and when that rating breaches a certain range, that path will
be shut down.

This means we won't be able to generate traces that would use the
discarded path, but when that path is not important it's a net benefit,
and when it was important CBMC wouldn't have completed anyway so we
trade an inaccurate trace for nothing at all.

Note: There is special consideration for loops: when any of a loops
children fails (inc. loops / recursion) it increments a counter, and
when that counter reaches a certain limit the whole loop itself gets
cancelled. This is done because when a branch deep within a loop gets
cancelled along all paths the loop will simply try again even though
it's been proven that the code will become too complicated to analyze.