Script to find regression tests that cover given source lines
An intended use case is to find regression tests that need to be adapted when a
loss of coverage in some files is observed when a new feature is merged.

For example, when improving the constant propagator to propagate more
operations, some existing tests that were intended to test the constraint
encoding of those operations might then be solved via constant propagation.
Thus, the existing tests need to be adapted to use non-constants as inputs.