Add underlying() methods on guard implementations
This allow to retrieve the underlying implementations used by guards
which is necessary if the user of guards wants access the guard in a way
that was not intended in the interface.
The usage we want to make of this is in the handling of guards to the
solver, we need to know whether the guard is an exprt or a bddt in order
to make this transformation as efficient as possible.