Add class_method_descriptor_exprt
This class stores information about a method on a class. Currently it is
only used by the java frontend, but there's no reason that it shouldn't
be used by others, like the C++ frontend. It uses ID_virtual because
that's what's currently used in the code, but it isn't just used for
virtual functions, so this will change in a future commit.