Add unit test for the string solver
This tests the solver directly on a set of desired formula instead of
through symex as would a regression test.
This can be easier to debug when trying to track a bug or add an new
feature, as the correspondence with the string solver code would be
easier to see.