Java frontend: encode array element types in classids
Previously java::array[reference] always carried the matching classid (i.e. "java::array[reference]").
Now it carries java::array[reference(Subtype)] for a single-dimensioned array of type Subtype, or
java::array[reference(array[reference(Subtype])]) for a two-dimensional array, and so on. This means
that instanceof tests for these types now work as expected, and ClassCastExceptions are thrown when
casting between array types (or corresponding ASSERT/ASSUME statements created when not running with
Java runtime exceptions enabled). remove_java_new and remove_instanceof passes are improved to deal
with this new dimension to the type system, and the array clone method propagates the type of the array
it is cloning as expected.

This should improve capacity to determine the real type of an array in a trace, and may help symex to
filter value-sets according to instanceof / checkcast instructions.

The remaining missing piece of the array type system is ArrayStoreException, as this requires us to
compare two unknown type IDs, cf. all existing instanceof operations which have a constant RHS operand.