Add new option --ignore-manifest-main-class
If the option is specified, any main class that may be specified in the
manifest file of a JAR will be ignored in the class loading stage.
This is useful when we want to e.g. load all classes in a JAR file, some
of which may not be referenced from the specified main class.