Do not call assign_from_json if not in symbol_table
The assign_from_json will fail with an invariant failure if it is called
on a method name that does not exist in the symbol table so we should be
careful when calling it.
This also make the function return early when the real clinit is not in
the symbol table, which can avoid some unecessary computation.