Optionally lift clinit calls after Java method conversion
This lifts all clinit calls in Java function bodies to the top, thereby
reducing the chances that a clinit method is repeatedly maybe-called (e.g.
due to `if(x) { A.staticfield.set(1); }`, which will only run A.<clinit> if
the branch is taken). If code like this occurs in a loop, for example, the
static initializer could be run by symex on every iteration, creating large
alias sets to account for all the possible initializer calls and sometimes
producing much larger formulas than are necessary.

On the other hand, with this option set, if `x` is definitely false we will
nontheless needlessly run the initializer, so in some circumstances the
option can increase costs.