Fix assigning to dynamically allocated structs
that is when

lhs :: [char]
rhs :: byte-extract-expr(struct, 0)

before we created a member expression at the offset and, in this case, assigned
to the first member of the struct. Now we assign the whole struct.

A test is included that was previously failing in the unwinding assertion (CBMC
did not see that the next pointer is eventually NULL).