Rename `component_name` to `mangled_method_name`
The component to which a `class_method_descriptor_exprt` is referring to
is specifically a particular overload of a method. The particular
overload is distinguished based on a mangling process. Therefore
`mangled_method_name` more accurately describes that this component is
the the mangled name of a method, as opposed to any other kind of class
component such as a field.