Disable GCC 8's class-memaccess warning so that MiniSAT builds
MiniSAT does a number of performance critical but unusual things
with memory allocation and layout.  These trigger a new warning
that has been added to GCC 8, class-memaccess and as we are using
-Werror, this means that CPROVER doesn't build using GCC 8.
I think it is better to disable this warning (which only seems to
be triggered in MiniSAT and by code that wouldn't pass review)
than to change the way MiniSAT works with memory.