Correct nondet init of function pointer arguments
At least until handling of function pointers is transferred to symex.

1. The nondet symbol factory only assigns nondet to function pointers, i.e. does
not build the `depth` deep pointer tree
2. The function pointer removal checks if the pointer is an argument to the
entry point and if so, skips adding the `assume(false)` for cutting short the
execution in case on function target was found.

add 2: On the GOTO level, the function call is replaced with `return_value`
which is never assigned to.