Add on_exit_scope_with_exceptiont
This is a scope guard that executes some code when a scope exits with an
exception. This is useful for performing ‘finally’ type actions without
having to wrap everything in a try block or rethrowing exceptions.

The header is called scope_guard.h because there are other kinds of
scope guards that I didn’t bother adding just yet because I didn’t need
them right now.