Report last useful source location on failure in symex
Sometimes we get failures in symex that contain diagnostics that are
useful for fixing bugs in cbmc, but not so much for telling users what
part of their code is causing problems during execution.
This reports the last symex step source location we’ve seen, which
should provide at least a decent hint of where to start looking for
problems most of the time.