Fix function contracts with ignored return value
This commit ensures that `--replace-call-with-contract` works for
calls to functions that mention `__CPROVER_return_value` in the
postcondition, where the call ignores the return value. Previously,
goto-instrument would not rewrite the `__CPROVER_return_value`
identifier in the body of the calling function, so CBMC would abort when
it looks up that symbol (which wasn't declared in the calling function).