Ensure string is non-null, radix is in range before parsing integers
Previously we passed the input to parseInt/parseLong straight to the built-in string solver functions,
meaning it was possible to crash jbmc by passing an out-of-range constant radix or for the solver to
produce crazy results when it regarded the radix as a free choice. Now the models apply the constraint
that it falls in a sensible range.

I add a test for parseint with a free radix, but am unable to add one for parselong yet as it takes
too long to solve.