String solver: support bounded-length intermediate strings, smaller index sets
This introduces two related features: the ability for the string solver to use constant-sized
arrays to store variable-length strings (which may be cheaper since smaller formulas are
generally needed to represent them and their constraints to the solver), and the ability to
maintain its array index-sets as sets of constants rather than symbolic expressions, which
can save a great deal of time and memory in the situation where there are many symbolic
indices that in all models occurring in practice tend to refer to the same indices.

Both are off-by-default for now since they are not heavily studied: I expect constant-sized
arrays would be expensive in the case where strings are numerous and sparsely constrained,
while storing index-sets as constants will result in more solver iterations when the solver
changes its valuation of symbolic index expressions each iteration (meaning the index set
grows despite the symbolic index set remaining the same size).