Allow old --export-function-local-symbols flag
This commit allows users to pass the --export-function-local-symbols
flag, which was deprecated in favour of the less-confusingly named
--export-file-local symbols. This was always the intention, but goto-cc
was erroring out when the old flag was passed. This commit instead
prints out a deprecation warning. This commit includes a test to ensure
that the warning is printed and that goto-cc successfully exports the
local symbols anyway.