Make assertions made by existing instrumenters customisable
This allows for additional logic to be added to the existing coverage
instrumenters. For example instead of generating an assertion of the
form `assert(guard)` we want to be able to generate an assertion of the
form `assert(!in_region_of_interest || guard)`. These customised forms
of instrumentation will still add assertions in the same locations and
be based on top of the original expressions.