on Windows, empty command-line arguments require quotes
Unix-like OSs pass an array of strings to processes as the command line
arguments. Empty arguments do not require special treatment.

By contrast, Windows uses a single string to pass all arguments to the
process.  The process parses the string to construct argv[].  To pass an
empty argument, quotes are required.

Note that this is unrelated to any additional parsing a shell might perform.