Tweak where we print the incremental status
We want to print the status after we've done any solving for that
unwind, but before we've done an extra step of symexing.

We should not print in the event the full equation has been generated,
as then we have a full solution so ins't incremental.

Ditto we don't need to print the status if we've found a failure, as
this function will be called again to do the next symex invocation if
there are further propreties to check