Java lambdas: accept direct method references
Usually an intermediate synthetic method is created by javac, but sometimes it can
use a direct reference to a method, which we would previously have failed to spot.

For example, Callable c = x -> x + 1 will generate a synthetic private method housing
the body of the lambda, but Callable c = SomeClass::someCallable may not generate one
but rather directly use SomeClass.someCallable as the lambda target. Whether this happens
in practice depends on the compiler.