Mark snapshot tests as KNOWNBUG
The C output doesn't quite work with the snapshot harness at the moment
(specifically, it doesn't work with the feature to restart analysis from
a specific point in the program) so these test are marked as KNOWNBUG
for now.

Going forward we may re-enable these after fixing the feature, or
possibly do without the restarting entirely.