Java frontend: support functional interfaces created by defining or explicitly abstracting methods
It's possible to make a functional interface by taking a parent or parents that have more than one
abstract method, then inheriting default definitions or making definitions yourself so that only one
abstract method remains, and so the interface becomes functional. Similarly you can re-declare a method
as abstract even if a parent made a default definition.

This commit adds support for all of these cases and tests them.