Java frontend: support lambda autoboxing against generic types
It's possible to use a primitive-typed method-reference to implement an appropriately
specialised generic functional interface (e.g. something that takes an integer can satisfy
Consumer<Byte> by direct method reference without needing to introduce a stub method to do
the conversion).

Here we implement that by unboxing using the java.lang.Number interface where necessary, thereby
catching all valid widening numeric casts at the same time (Short.intValue for example will
do the widening and deliver the required primitive type). When boxing we always box to the
primitive type's corresponding boxed type, as this is all the compiler seems to allow (i.e. it
won't allow satisfying Producer<Long> using a function that returns an int, even though that is
type-safe).

Note this solution uses virtual dispatch a little more often than is needed: in some simple cases
the class file contains enough information to know that the Object we're unboxing must in practice
be an Integer (for example) and so no dispatch is necessary. That doesn't account for all cases
where a generic interface is implemented however, so I did not pursue this possibility.