Make filter_goto_model remove file local symbols
Since we want goto_harness to print out as little as possible to be
useful this is helpful, and this also lets us avoid bugs with duplicate
definitions (e.g. a static variable that’s now in the produced harness.c
– and thus a *different* static variable!).

Also marks non-file-local globals as extern so we avoid double
definitions.