Rather than change existing tests, create new test descriptions
A number of the tests for the old constants domain had been
converted to test the variable sensitivity domain.  This would
reduce the test coverage of the --constants domain.  So instead
new .desc files have been added so both can be tested on the
same input programs, allowing side-by-side comparison.