Add test of binary inclusion in xml trace for integer values
In order to confirm that the binary is now included in the xml trace
output.

Lines may be separated with `\r\n` on windows, rather than just `\n`. So
we need to match both kinds of separator in order to be platform
independant. As we don't require the xml elements to be on separate
lines and `\s` matches both `\r` and `\n` we can let `\s*` match all the
line separators as well as any indentation characters.

`uint32_t` for example fully specifies that an integer is unsigned and
32 bits. But it could be implemented as a `typedef unsigned` or a
`typedef unsigned long`, so long as these underlying types fulfill the
requirements. Therefore the regex for this needs to match both `16u`
and `16ul` as the value.