Add utility modules for byte extract and update metric reporting
Provides the utility modules required to report byte extract
and byte update metrics. This includes helper functions to
check if given expr is a byte extract or update and an SSA_stept
class member function that returns the relevant ssa expression
from a given SSA step. If the SSA step is a shared_read or
shared_write, return ssa_lhs, else return cond_expr.

Also, added an invariant in the constructor for byte_extract_exprt
and byte_update_exprt to confirm _id is valid for additional
correctness checks.