Create a new abstract interpreter that does a three-way merge
Three-way merging is a VSD specific feature.  On function call return,
rather than just merging in the final state of the function to the
return context, it computes the difference between the first and last
states and then merges that into the calling state.  The key difference
is that if there are multiple calling contexts, their state will be
merged and normally picked up.  For example;

global = 1;                  global = 2;
f00();                       f00();

The start of f00 must have global as the merge of the two (i.e. top if
constant) and this will be merged into the return state -- even if
f00 obviously does not touch global.  Three-way merge only modifies
the variables that change during f00 so in this case, global will be
unmodified in both contexts.

All of this only makes sense if there are multiple calling locations that
merge into the same history.  So if a fully context sensitive history
is used then there is no point using three-way-merge.

This commit also updates the tests that actually need this functionality.
They demonstrate some of the other advantages of this technique.