Add parse option --list-memop-calls to tracks calls to memcmp, memcpy and memset functions
Memory operations often increase proof runtime and need to be replaced by stubs
for improved performance. Here, we track all calls to memcmp, memcpy and memset
(can be extended to other memory operations) for ease of identification and any
subsequent stub replacements.
This commit adds a parse option in CBMC to achieve this.