Add missing perl command in regressions/CMakeLists.txt
We were facing an issue in Windows, where the shebang in
`test.pl` was not interpreted properly, causing the tests
not to run reliably.

It could be possible to hack something similar together,
by associating file endings with runners which is what
python does, but this doesn't work for perl.