Move ${gcc_only} outside of quoted strings.
Fix its definition, and make it be the last argument
passed to add_test_pl_profile om cbmc/CMakeLists.txt

The reason for this is that in CMake, a "string" is
a list of one element, and a "list" is a string with
semicolon separated items.

In its previous format, whenever ${gcc_only} was expanded,
we would get something like "-X;KNOWNBUG;-X gcc-only",
which is wrong because it's a single argument that happens
to contain spaces. Passed to `test.pl` like this, it would
fail to work, because `test.pl` doesn't know what that is.

Another possible transformation that would work, highlighting
the "lists are strings and strings are lists" in CMake is this:

set(gcc_only "-X;gcc-only");