Add goto-unwind skeleton
This is supposed to be a tool that ‘unwinds’ goto-programs, by unwinding loops,
inlining and following function calls (including recursively) from some starting
point, similar to what CBMC does in its symex phase but producing a goto-program
instead of an SSA program.

For now this is just a skeleton that only does loop-unwinding like the existing
`--unwind` (and related) flags in goto-instrument, but these don’t quite fit our
needs on this and so we decided that putting this funcitonality into a separate
tool would be easier to manage (rather than adding even more configuration
options to the already fairly crowded `goto-instrument`).