xen-cbmc: fix ci script
This script has been broken, and not been picked up in CI until now.

This change fixes the script, as well as makes sure that the relevant
steps are executed, and the resulting build artefacts have the required
goto-cc section in them.

This CI script has been refered to as part of the paper:
"Using model checking tools to triage the severity of security bugs in
the Xen hypervisor", that has been presented at FMCAD 2020.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>