Move "is hidden" attribute of goto functions out of type
Whether the execution of a goto function is to be hidden has nothing to
do with its signature, and therefore should not be included in the type
(irrespective of whether the type is stored in goto_functiont or in the
symbol table).

As this attribute is recomputed upon reading a goto binary, there is no
need to store it in the goto binary, and thus also no need to bump the
goto-binary version number.