Reorder address checks
This is a slight change in functionality. A future refactor of the
address checks will necessitate reordering these anyway, and doing it in
isolation first gives us some additional confidence that our refactor
didn’t have any unintended consequences aside from the reordering.

Also changes the regression tests that break because of the reorder.