Concurrency: ensure apply_condition's value is used
goto_statet::apply_condition generates an entry (with fresh L2 index) in
the propagation map with generating an assignment. We did not previously
use the propagation map when handling atomic sections, and thus had a
missing link between values, causing spurious verification failures.

The regression test is based on SV-COMP's
pthread-ext/23_lu-fig2.fixed.c.

There is further room for improvement, as shown in the
KNOWNBUG-regression test. This will be addressed in a separate commit.