Fix bug with function pointer initialisation in presence of statics
Previously the code used some logic to skip generating selector cases
for static functions, which is what we want because we can't see
functions in the harness. Problem is if that happens to be the *last*
function we skip generating the "else" case at the end, so we end up
leaving the function pointer potentially uninitialised.

This just removes the functions we want to skip from the targets list.
Bonus: Loops without skips are easier to read.