Restore homebrew PR submitting after new release.
Remove temporary blocker of execution for the
github action responsible to submit a PR to the
homebrew repository, updating the CBMC version there.

This was necessary during the testing of the
dockerhub image push action, as we needed to test
it live, and we didn't want to spam the homebrew
repository with PRs pointing to testing/experimental
PRs of ours.