Fix doxygen in cmake
This has been broken for 3 years, ever since doxygen.cfg was renamed to
doxyfile. This lets us build documentation in from cmake with the 'doc'
target. One difference to the previous version (i.e. the one that was
broken before) is that this creates the documentation in the cmake
binary directory now instead of the source directory. Reasoning for this
is that cmake targets should not change files in the source directory
ideally.