Remove {f,F}orall_subtypes
This was useful in the past, but with C++-11 we can use a ranged-for to
avoid the iterator altogether.

As subtypes are no longer stored in a named sub (removing the
corresponding #if-0-disabled code), the ranged-for can be used directly
without testing for the existence of subtypes beforehand, as the forall
macro did.