Enable the CPROVER library for goto-analyzer
This does change the functionality quite substantially but strangely it
only seems to affect one test, and that is syntactic rather than semantic.
As with many of the other changes in this PR, I suspect this is swapping
one lot of broken, untested functionality for another, slightly less broken
but still not working lot of functionality.