Add string abstraction to goto-analyzer and goto-diff
At the moment this code is dead and is only added to make the
process_goto_program steps more similar.  I will leave it to others
to decide if this is a useful thing to have and add the required command
line options and tests.  Should not affect the behaviour of either tool.