Array-element sensitive SSA: avoid string-to-integer-to-string
There is no need to use decimal indices in field-sensitive SSA
identifiers: we just need pairwise distinct names. Thus, just use the
irep-stored value instead of converting the value to an integer and back
to a string.

On one (array-heavy) benchmark, this reduced the time constructing the
symex target equation from 74 to 60 seconds.