Add overview of the CProver tools
This adds the TOOLS_OVERVIEW.md file (and links to it
from README.md). The Tools Overview is an overview of all
the tools that are part of the CProver package. Each
overview includes main points on how to use the tool and
what the tool is meant to do. Also further documentation
links are provided to better understand the tool.