Make the call to rewrite-union configurable
Rewrite union converts union member access to byte extracts.  This is
(apparently) good for CBMC and goto-diff but it makes analysis harder
for goto-analyzer.  So to make them consistent we add an option which
controls whether it is run or not.  This is fixed in each tool but it
takes us closer to common processing of goto programs.