Add options that goto-check uses to the macros
goto-check makes use of the options assertions, assumptions,
assert-to-assume and retain-trivial.  These were handled to different
degrees and in different way in each of the tools.  This commit adds
these options to the set of options in the *_GOTO_CHECK macros.  This
reduces the amount of duplication and parallel maintenance.