goto-harness: create goto binary or C output
It is not always necessary or desirable to produce C output. An ability
to produce goto binaries also avoids possible limitations of C and/or
dump-c (such as goto-harness creating initialisers for structs with bit
fields, which yield functions that take bit fields as parameters, which
isn't possible in C).

While at it, also clean up the help output to fix confusing language
about goto binaries when the output was actually C code.

Fixes: #5351