Add no initialisation PRECONDITION to `wp_decl`
This function does not account for the case where a code_declt contains
`initialization`. Given that the test suite passes with this
PRECONDITION is in place, such statements must be separated out before
this function is reached. Should this condition be violated in future,
then the CONDITION can be removed and the handling of this case
implemented.