Pulled out interval_evaluator
Wrap up interval evaluation in a wrapper class.

Most significant change is simplifying the process of
coercing constants to intervals. Rather than trying to bail out early if
we hit a top value or something which can't be converted, go over the
whole list of operands, skipping any we can't convert. Check afterwards
if we're missing some and then decide what to do.