export-file-local-symbols: handling files with many functions
The use of --export-file-local-symbols entails renaming of file-local
(function) symbols with a prefix of __CPROVER_file_local, which had the
side effect of those names being added as defines when building hybrid
binaries. For files with large numbers of static functions, this
resulted in a command-line that exceeds operating-system limits.