goto-instrument-wmm: removing graph edges invalidates the iterator
Store the elements to be removed in a set until after the iteration is
done, and then remove the chosen elements. The more standard approach
could be returning the iterator to the next element in the
remove-an-element function, but this requires larger changes here. Thus
keeping it small and simple for now.