Remove fixed windows SDK version lookup.
Before this, we were looking for the codesign tool under a specific
folder, signifying a specific version of the Windows SDK. But Github
Action Runner Software is being updated every now and then, causing
this dependency to be wobbly.

This should make the tool lookup more robust for future SDK version
updates as well.