Add abstract_value_objectt::to_interval()
Analogous to to_constant, it returns a constant_interval_exprt that
expresses the range of the value. For constants the interval is across
whatever the constant is, for intervals it is their interval(!) and for
value_sets it is the interval between the smallest and largest values in
the set.

Used to intervals to merge with any other value type.

This commit includes unit tests for interval_abstract_objectt::merge.