Log version of z3 installed in CI
For confirmation of version and that installation is successfull. This
supports easier debugging for both the github actions `.yaml` and for
future test failures. As a bonus we get fast failing CI jobs if the
dependency isn't installed properly, because it fails at this step
rather than during testing.

This z3 version check is done as a separate step because adding to
`$env:GITHUB_PATH` doesn't take effect the path until the next step.