Use `Invoke-WebRequest` instead of `choco` to install `z3` on Windows
Because the `choco` installation of `z3` seems to be broken/outdated.

Using `Invoke-WebRequest` instead of `choco` allows us to pin the
version installed and to get a version which includes the `z3.exe`, not
just `z3.dll`. Using `nuget` was also considered, so that we were still
using a package manager. But the up to date `nuget` package was missing
the `.exe` file.

`wget` is supposed to be aliased to `Invoke-WebRequest` in powershell.
However `Invoke-WebRequest` is used rather than `wget` as `wget` does
not seem to work for windows github actions jobs.