Manage lifetime of Glucose solver pointer using `unique_ptr`
This has the benefit of allowing the use of a default destructor rather
than a hand written destructor. Note that `satcheck_glucose_baset` was
already non-copyable because its `solver_hardness` field was
non-copyable.

The default implementation of the destructor is in the `.cpp` rather
than the header file in order to avoid recompiling it every time the
header file is included and so that the declaration of the solver
destructor is available when instantiating the destructor template of
`unique_ptr`.