Deduplicate output streams in `smt2_convt::write_footer()`
This fixes bugs arising from some of the member functions of
`smt2::convt` writing to an `os` argument and some of the them writing
to `smt2_convt::out`. This was causing some of the generated text to be
written out when using `--outfile` argument with cbmc, but not without
it.

It might be easier to understand `smt2_convt` if the output stream was
passed as an argument to the member functions, instead of kept as a
member variable. This commit replaces the arguments with references to
the member variable because it is a less wide sweeping change and it
still solves the bugs.

The address_space_size_limit3 test is disabled with smt2, because the
output of this test was already different between the sat and smt2
backends and therefore it was already not producing the correct output.
Presumably this test was returning the expected `VERIFICATION FAILED`
result for the wrong reasons. The `VERIFICATION FAILED` result is only
returned in the case where the object size limits are missing. Further
work will be needed to properly debug the support for `--pointer-check`
with smt2.