Fix for iterators into unordered_maps
C++11 §23.2.5/13 states:

The insert and emplace members shall not affect the validity of references
to container elements, but may invalidate all iterators to the container.

This fixes two instances where the iterators are assumed to remain valid
while inserting additional elements into an unordered_map.  The problem is
exposed when using -D_GLIBCXX_DEBUG.