Add handling of "unknown" SMT outcome
Currently it is possible for z3 (and maybe other SMT solvers) to
return a satisfiability result of "unknown". At the moment this is
treated as an ERROR outcome in cbmc. This commit changes the
error message handling to output why the result was ERROR by
reporting the "unknown" satisfiability result.