Add conversion of bit_vector literals to smt terms
The previous implementation duplicated the type to sort conversion from
`convert_type_to_smt_sort`, in checking the type of the expression. By
converting the type to a sort and then specialising based on the sort
rather than the type, this duplication is eliminated.
`convert_type_to_smt_sort` now defines the kind of encoding used for any
given expression.