export-file-local_symbols:: update base name and pretty name
Name mangling via --export-file-local-symbols used to introduce symbols
the name of which did not match the base name. Forward declarations of
the mangled name, however, did introduce a symbol where (mangled) name
and base name match. When linking later found the definition
corresponding to such a forward declaration, the base name (and pretty
name) were not updated. As a consequence the order of linking produced
two different final symbol tables.

Note that Crangler may be a superior alternative avoiding such problems,
but there is ample code that still uses `--export-file-local-symbols`.