Recanonicalize with_exprt when renaming old member
It may happen that during renaming the `old` subexpression of a `with_exprt`
expression is propagated with a value that has an array type with a size that
is a symbol with an L2 index that is different. In this case the type of the
`with_exprt` will not match with the type of the `old` subexpression
anymore.
To address this issue we re-canonicalize the `with_exprt` by propagating
the type of the `old` subexpression to the type of the `with_exprt`.