Fix printing of `smt_get_value_commandt`
The get value command is supposed to be parameterised with a collection
of one or more terms for which the solver should return the values. Our
internal representation of the get-value command only supports a single
term, which should be placed in brackets to make it into a collection
containing a single term. The brackets were previously missing, which
would have caused a solver error.