fix a bug with inclusion checking on havoc_object
The write set inclusion check on havoc_object(ptr) only checked the
inclusion of (ptr+0) in the write set, but performed havocing of the
entire object (i.e., ptr+0 until ptr+OBJECT_SIZE(ptr)).
The proposed change fixes this behavior and checks the inclusion of the
entire object in the write set.