Support byte update lowering of/from an array of bits
Byte operator lowering made several assumptions about array elements
being byte aligned, which may be true for ANSI C, but isn't the case for
our C front-end (which supports arrays of single bits), and not true for
the overall framework in general.

A key part of this is factoring out code that was almost-the-same from
lower_byte_update_array_vector and lower_byte_update_struct.