bv_typet is also sensitive to endianness
The bits within an object of bv_typet are ordered depending on
endianness. This is consistent with typecasts to/from bv_typet of some
other bitvector type, which do not alter the sequence of bits.

The unit test would previously fail the check `REQUIRE(lower_be1 ==
*array_of_bits)` for the big-endian case (line 89), even though the
little-endian case was fine.