Invoke goto_functions.update() after inlining
Inlining will alter location (instruction) numbers, which need to be
updated to once again be globally unique.

With this change, invar_check_large requires an assigns clause for the
loop as the local may alias analysis returns "UNKNOWN." The previous
result was actually unsound, because the automatically computed assigns
set did not include arr0, arr1, arr2, arr3, arr4. This also demonstrated
that the implementation in the test was incorrect: it would not terminate when:
*(arr[h]) == pivot && *(arr[l]) == pivot && h != r && l != r.