Wrap NULL/0 is-a-null-pointer in a utility function
A pointer-typed constant expression with value "NULL" is always a null
pointer expression. When NULL_is_zero is true in the platform
configuration, a value "0" is also a null pointer.

These possible options were repeatedly, but not consistently, evaluated
ad-hoc. This resulted in a least one missed simplification opportunity
as we noticed on an RMC-generated goto binary (which had 0 instead of
NULL in a pointer-typed constant).