Function contracts: performance optimisation for assigns clause checking. - don't instrument assignments to locals and function parameters, - don't add function parameters and non-dirty locals to the write set, - remove from the local write set CARs that are provably not alive, - remove function parameters and locals from assigns clauses in tests.
Sanity checks:
- change GOTO instruction with false guards into SKIP instructions
  (removes artificial loops),
- check loop-freeness before assigns clause instrumentation
  (required for soundness of assigns clause checking).

The net effect of these changes is a better proof performance because
of a reduction in the number of generated assertions and of their
size, but otherwise the contract checking functionality remains
unchanged.

Rationale:

Assigning to local variables or function parameter is always
legal so we skip instrumenting these assignments. We avoid adding
function parameters and local variables to the local write set, except
when their address is taken at some point in the program and can later
be assigned to indirectly via pointers.

When generating inclusion checks for assignments, we remove from the
local write set targets which are not possibly alive at the ASSIGN
instruction that gets instrumented.