Improving linker warning diagnostics
We previously just reported pointer type conflicts without elaborating
what the exact difference was. To produce type-consistent goto programs,
we chose to use the type that the definition had, but lacked any test
that doing so would actually result in a program that can be analysed
successfully.

This commit fixes bugs in the code producing diagnostics, enables
diagnostics for the case of pointer type conflicts, and adds a test to
demonstrate that we successfully analyse programs after type conflict
resolution. Despite this, however, inconsistencies in the goto program
remain: the call types still need fixing up.

Fixes: #198