Consistently use exprt::is_constant()
We would sometimes use is_constant and at other times id() ==
ID_constant (which is exactly what is_constant() does). One option was
to get rid of is_constant(), but it seemed safer to use this function
than to stick with the lower-level id() == ID_constant: the function is
type safer as it can only be used on exprt, while comparisons on id()
would work on any irept.

To ensure there is no performance penalty, is_constant() has been moved
to the header file to permit inlining.