Consistently use exprt::is_boolean()
We would sometimes use is_boolean and at other times type().id() ==
ID_bool (which is exactly what is_boolean() does). One option was to get
rid of is_boolean(), but it seemed safer to use this function than to
stick with the lower-level type().id() == ID_bool: the function is type
safer as it can only be used on exprt, while comparisons on id() would
work on any irept.

To ensure there is no performance penalty, is_boolean() has been moved
to the header file to permit inlining.