Regression test scripts: use /Fe with goto-cl
Instead of a build-then-move dance, just use the appropriate
command-line flags. Some scripts already did this, now all of them do
so.

For goto-instrument/chain.sh that includes a more substantial cleanup
that removed a chunk of dead code ([[ x$name == x ]] was necessarily
false). Also, this script is no longer hard-coded to create main.gb as
output so as to permit multiple tests in the same directory (with
different target files) that could potentially be run concurrently.