Make constant_exprt a nullary_exprt
It should always have been one, and the introduction of
annotated_pointer_constant_exprt makes it possible to implement this.

This highlighted to pieces of dead code: there are not `constant_exprt`
of type `ID_array` for such constant are `array_exprt`;
`goto_trace_constant_pointer_exprt` was never used (though meant to be
what annotated_pointer_constant_exprt now actually is).