Refactor usage of `PeekNamedPipe` to get count rather than read
This makes it more obvious from reading the code that no characters
will be removed from the pipe as part of the implementation of
`can_receive`.

This commit includes renaming of local variables to exactly match the
names and types in the API definition, with nullptr used to distinguish
pointers from integers.