Bugfix communication by keeping original input pipe in parent process
Prior to this commit, communication sent to the solver was not being
received. Sending the data was reported as successful, but the
subprocess could not read it from the standard input.

The inherit flag is now set rather than cleared, so that the sub
process is able to use the handle.

This commit enables all piped_process unit tests on Windows, as these
now work/pass.