Add INVARIANT on the expected number of bytes being correctly written
This could end up mismatching for sufficently large messages exceding
the length of the pipe's buffer. For the moment it is better to halt if
this potential issue is encountered rather than risk introducing
communication problems. If this problem is encountered later then it may
be possible to work around by splitting the message into multiple writes
to the pipe.