C library: move pthreads related definitions to library
With the support of building initialisers while adding the library we
can move library-only symbols to the library. This avoids polluting the
initialisation function for all cases that don't actually use
pthreads-related library functions.

This removal of unused symbols required adding a symbol into the
array_of_bool_as_bitvec test, which previously implicitly relied on
those library-specific symbols in the patterns being tested for.