Single-path symex must not fail an invariant when verification succeeds
Single-path symex takes a best-effort approach to produce a proof by
using the symex equation of that path that was last visited. This,
however, failed an invariant when trying to access the equation that had
already been popped from the worklist. Leave the last equation in place
to avoid this problem.