CONTRACTS: new method to detect static locals for contract instrumentation
The previous method relied on building a call graph and
could not be used on functions that had already been through
an inlining pass.

This new method uses source_locations and manages
scoping by tracking the min and max source_locations for
functions that are encountered when traversing instructions.

Then, when collecting statics from the symbol table, we only keep
the ones that are found in the traversed functions and whose
source_location is within the seen intervals.

The method makes the sound pessimistic choice to widen the tracked
intervals and collect variables whenever the encountered
source_locations are undefined or only partially defined.

When replace-call-with-contract is used, extra pragmas are used
to track static locals and make sure they are propagated to the write set
of the function in which the call is replaced.