Remove Throwing-NULL check from goto_check_ct
This check was introduced in d2713b200, with no particular reason that
it should be enabled unconditionally. The test included in that commit
has --pointer-check set, suggesting that this always was the desired
behaviour. It was, however, only intended for the Java front-end: for
the C language, "throw" is not relevant, and C++ permits throwing NULL.
Therefore, remove this check from goto_check_ct.