Python script to fetch declarations of Clang's intrinsics
Thus far, we constructed most of the declarations using the
get-gcc-builtins.sh script. We may need to continue to do so for GCC.

Clang's declarations, however, seem easier to parse. This new script now
takes care of this.

In future, we may decide to completely remove the GCC-specific script,
and build upon Clang's declarations with selective tweaks instead.