goto_rw: do not assume that all types have constant size
pointer_offset_bits returns nullopt whenever the size of a type is not a
numeric constant. Such cases arise when arrays have runtime-determined
size, as is the case in regression test byte_update5.

All other places except the one fixed in this commit already considered
the case that pointer_offset_bits returns nullopt.