Implement extract_bits conversion
This only works so long as the bits to extract are constant values. This
is the same limitation as is present in the old back end. The reason for
this is limitation is due to the way the `extract` operation is defined
in the smtlib version of bitvector theory. It may be possible to work
around this limitation but the work around is unlikely to be trivial.