New built-in __CPROVER_pure_contract.
This built-in allows to tag a function body as a pure contract.  It
can only be used as the sole function call in a function
body. Incorrect uses result in conversion errors.  It gets converted
into assert(false) with property class ID_pure_contract to force the
user to replace calls to such functions by their contracts.