Document and use malloc(0) behaviour
realloc(ptr, 0) can return an object of minimal size. We used to use a
size of 1, but there does not seem to be any guarantee in the standard
that "minimal" refers to a value greater than 0, so use that size
instead. Using malloc(0) is safe with our library model, and is now also
documented.

While at it, also move this documentation to a place that has more
context.

Fixes: #6360