Goto-symex: record value-at-declaration in the back-end
Declarations are printed as part of a symex trace, and come with both a
symbol being declared as well as an (unconstrained) initial value for
that symbol. The decision procedure must learn about the initial value
to include it in the model. Else, the trace value would just be question
marks.

Fixes: #6845