Scripted parallel property checking using CBMC
Adds a (Python) script to check all properties contained in a given GOTO
binary individually, but in parallel. This can yield a somewhat
different user experience in that some of the results become available
early on while other properties are still being checked for, as opposed
to the all-or-nothing experience when doing all properties at once.