Remove unused option "simplify-if"
Going back as far as version 2.9 of CBMC this option does not seem to be
read anywhere, it was just (unnecessarily) being set in the
front-end(s).

The command-line option "no-simplify-if" is left in place so as not to
break any scripts. A major version upgrade should take care of cleaning
this up as well.