Update CBMC man page
With command-line help being complete, help2man was used to create an
updated list of options for CBMC's man page. Moving forward, automated
checks for man-page completeness will be put in place so that future
changes to the command line will have to be reflected as part of the
changing commit.