CONTRACTS: detect and handle locals in assigns clause checking for loops
When no assigns clause was provided, we were trying to infer variables
touched by the loop and generate an assigns clause from that.
Some of these variables could be locals, which have no meaning
outside the loop, and we were trying to snapshot these variables outside
of the loop. We now detect loop locals and remove them from the inferred
frame condition the loop.