Byte update lowering: never create extractbits over unbounded arrays
With nondet pointers we may end up doing byte updates of weird
combinations, including doing byte updates of an unbounded array using a
value the size of which is not a multiple of byte widths. This resulted
in extractbits expressions trying to extract bits from an unbounded
array, which bitvector flattening does not support.

This commit puts a byte extract layer in between, which will then take
the right approach to ultimately enable the extractbits to succeed.

Also, put PRECONDITIONs in place to ensure we don't create such
extractbits expressions in any other place in byte-operator lowering.