Use cmdlinet::get_comma_separated_values where possible
Ensure that we accept repeat use of command-line options the
arguments of which are comma-separated values. That is, allow command
line options that take values consisting of comma separated lists to be
repeated and accumulated into a single list. For instance, making
`--foo A,B,C,D` be semantically equivalent to `--foo A,B --foo C,D`