Remove well-alignedness assumption from pointer dereferencing
With void* pointers, pointer dereferencing would permit any type for the
dereferenced object. For arrays, alignment at the root level was
implicitly assumed, not taking into account that there may be a
well-aligned pointer to a nested element.

For the regression test we got `arr[11 / 7]` when dereferencing `(void
*)ptr`, where `ptr` was `(char *)&arr[0] + 11` (and the size of `arr`'s
elements is 7 bytes).

Fixes: #7036