CONTRACTS: Filter out contract symbols when doing lookups by base name.
Contract symbols (#6799) have the same base name as the function symbol
they are derived from. This causes both the function and its contract
to be found when doing a lookup by base name, when resolving entry points
for instance.

We now filter out symbols that have the `is_property` set when resolving
entry points and interrupt handlers in a goto model.