Enable compact less-than propositional encoding
This was previously disabled as it appeared to degrade performance. New
benchmarking, however, suggests considerable performance improvement:

When running on various proofs done for AWS open-source projects, this
changes the performance as follows (when comparing to #7021): with
CaDiCaL as back-end, the total solver time for the hardest 46 proofs
changes from 26779.7 to 24472.6 seconds (2307.1 seconds speed-up); with
Minisat the hardest 49 proofs take 18541.2 instead of 28420.4
seconds (9879.2 seconds speed-up). Across these benchmarks, 1.0% of
variables and 3.2% of clauses are removed.

INCLUDE_REDUNDANT_CLAUSES is not enabled: while this would yield a further
speed-up of 1080.4 seconds with CaDiCaL, it slows down Minisat by 4440.6
seconds on the above benchmark set.