CONTRACTS: Add the `dfcc_dsl_contract_functionst` class.
The class translates the assigns and frees clauses of
a contract expressed in DSL style into goto functions
returning `__CPROVER_assignable_t` and `__CPROVER_freeable_t`
types and describing the same set of targets as the clause.

Internally it uses the `dfcc_spec_functiontst` class
to translate the functions generated from the clauses
into functions that dynamically build write sets at
runtime.

A havoc function is also generated for the assigns clause,
that havocs the locations described by the clause in
a type-directed way, taking snapshots of start addresses
from a write set instance.