Add non-determinism to array_address_of test
This test is intended to cover incremental SMT decision procedure
support expressions of the form `address_of(array[0])`. These occur as
part of the conversion from array to pointer, which may be implicitly
introduced by applying `__CPROVER_OBJECT_SIZE` to an array. The existing
test is not actually checking this conversion due to it being simplified
away during constant propagation. The introduction of non-determinism in
this commit prevents constant propagation from removing expressions of
this form, so that their conversion will actually be tested.