Allow characters with value >= 128 in XML output
These values are used as part of printable characters which should be
allowed in XML output. This invariant was previously violated for
characters such as `¬` and `⇔` which are included in the output when
cbmc is invoked with `--verbosity 10 --xml-ui`. This was due to the
comparison formerly being a signed comparison. So byte values in the
range 128 to 255 were treated as values in the -128 to -1 range which
are less than ` ` which has value 32.