CONTRACTS: Avoid single conjunction in contracts instrumentation
Currently, requires and ensures clauses are always combined into
a single conjunction. The resulting clause is then used to create
a single assertion or assumption, which makes really difficult
to debug issues with large function contracts. This commit
preserves the source location of the clauses by creating one
assertion/assumption for each clause in the function contract
(i.e., either requires or ensures).

Signed-off-by: Felipe R. Monteiro <felisous@amazon.com>