Value sets: type casts need not be towards a pointer type
We cannot assume that the expression seen by get_value_set_rec is of
pointer type. Therefore, type cast expressions need not be towards a
pointer type, and unconditionally trying to take the subtype of the
expression type is not safe. On the other hand, any access to
NULL_object will be invalid, and fixing NULL_object to be void typed
should be safe.