C library: fscanf for modern glibc
Recent versions of glibc rename fscanf (and also some others) to ones
with a prefix of __isoc99_, rendering our previous C library model
useless. This commit adds a copy of the existing model with that new
name, and also seeks to use it in a regression test.

The original goal, however, was to exercise a piece of code in string
abstraction.  Exercising this code now works, but results in an
invariant failure. The test, therefore, remains marked KNOWNBUG.