Remove unnecessary includes
Manual removal of includes based on include-what-you-use's output,
filtered for includes that should be removed. The goal is to avoid
unnecessary build dependencies, reducing the amount of code that needs
to be rebuilt during incremental builds.

This is a second iteration of this cleanup; the first round took place in
450845dbc5.