Field sensitivity: using apply is now always safe
With the changes from 619353432 it is safe to call `apply` even when
doing field assignments: the right-hand side will not be a divisible
compound (which we previously sought to protect from field expansion).

This makes field sensitivity stateless, and all methods are `const`.
They are not static, however, as the constructor initialises
configuration options (which would otherwise need to be passed to each
call).

We can then also avoid a level of indirection through `apply` and call
`get_fields` directly in two cases.