Fix printing of SMT data structures in failed unit tests
The operator needed for the printing is defined in
`smt_to_smt2_string.cpp`, but it needs to be forward declared for the
catch framework to find and use it, instead of printing SMT data
structures as `{?}`.

A regression test of a failing unit test is included in this PR to
ensure that this functionality for fault finding of failing unit tests
works as intended.