[CI] Fix windows slack notification.
We currently have a small `go` utility to produce some notifications on
Diffblue's Open Source team channel on Slack, so that we keep track
of the release and any builds for associated artifacts.

The notifications work on Linux, but they required running the `go`
program directly on MacOS, because the action is bootstrapping an
Alpine Linux container which for some reason is incompatible with
the MacOS and Windows runners.

This commit follows the same recipe (running the go binary directly)
to allow us to keep track of the Windows `.msi` installer package
build.