CONTRACTS: add __CPROVER_obeys_contract predicate
This predicate can be used in requires and ensures
clauses to express that a function pointer satisfies
a given contract.
The predicate is implemented in cprover_contracts.c
and turns into an assignment in assumption contexts
and to an equality check in assertions contexts.
The dfcc_instrumentt methods were updated to report
any occurrence of the predicate discovered when
instrumenting GOTO functions.