Remove internal symbols: keep all symbols used in contracts
In beb1353cf2 we had chosen to only keep code-typed symbols, but indeed
we also need to retain objects (only) referred to from contracts.

Also add debug-level status information of what symbols are being
removed (which is information I have previously needed more than once in
debugging, and it's also now the only actual use of the messaget that we
already had in place).

Fixes: #7414