Byte operator lowering: no type casts to/from compound types
We need to create proper struct or union expressions instead of doing a
type cast, even when the compound type is no wider than a single byte.
Equally, we must not fall back to just type casts to bit vectors when
there is an array of structs/unions where those are exactly one byte
wide.

Fixes: #7288