Incremental SMT back-end no longer requires --slice-formula
With #6590, several unnecessary symbols are no longer included in goto
programs. These include unbounded arrays, which the incremental SMT
back-end hitherto does not support. The use of --slice-formula was a
workaround to make sure the unnecessary symbols (or equalities over
them) don't end up in the formula; now they aren't part of the goto
program, so the workaround no longer adds value.