Field sensitivity: fully expand nested objects
When applying field sensitivity to an expression we must expand all
levels. In the included test we previously constructed a member symbol
(`var_11..value`) as right-hand side that still was of struct type.
Instead, we must create a struct expression composed of all its members
(here: `{ var_11..value.._0 }`).

Also, we must not keep nested expandable members in the propagation map
for phi nodes don't merge on them.

Fixes: #7462