Update base Ubuntu image to 22.04 from 20.04 in `Dockerfile`.
This allows us to tap into the udpated software registry that
Ubuntu 22.04 uses, which allows us to use a new CMake version (`3.22`),
which we want to be able to build the Rust API, for which we depend
on a package that needs `CMake 3.19`.

It's worth mentioning that the rest of our CI infrastructure is already
at `3.25`, and personal development machines should be between `3.22`
and `3.25-3.26` (for MacOS machines using `brew`).