Remove @pkesseli to clean up github errors for the codeowner's file.
@pkesseli no longer has write access to the repository (as his access
was inherited through his membership to the @diffblue organisation)
so now Github is throwing some errors claiming that the CODEOWNERS
file is invalid.

This is removing Pascal so that the CODEOWNERS file can be free of
errors again, but we can reinstate Pascal if he desires to maintain
his codeownership through other means.